+++ src/cacheout/lfu.py
@@ -34,15 +34,22 @@ class LFUCache(Cache):
         # keys first (i.e. keys with a higher count).
         self._access_counts[key] -= 1
 
-    def _get(self, key: t.Hashable, default: t.Any = None) -> t.Any:
-        value = super()._get(key, default=default)
-        if key in self:
-            self._touch(key)
-        return value
+    def get(self, key: t.Hashable, default: t.Any = None) -> t.Any:
+        with self._lock:
+            value = super().get(key, default=default)
+            if key in self._cache:
+                self._touch(key)
+            return value
 
-    def _set(self, key: t.Hashable, value: t.Any, ttl: t.Optional[T_TTL] = None) -> None:
-        super()._set(key, value, ttl=ttl)
-        self._touch(key)
+    def set(self, key: t.Hashable, value: t.Any, ttl: t.Optional[T_TTL] = None) -> None:
+        with self._lock:
+            super().set(key, value, ttl=ttl)
+            self._touch(key)
+
+    def add(self, key: t.Hashable, value: t.Any, ttl: t.Optional[T_TTL] = None) -> None:
+        with self._lock:
+            super().add(key, value, ttl=ttl)
+            self._touch(key)
 
     def _delete(self, key: t.Hashable) -> int:
         count = super()._delete(key)